[Recording API] Trying out 16bit numbers

+++ inc/MicRecorder.h
@@ -44,7 +44,9 @@ namespace codal{
         int             position;                       // Pointer to a position in the savedRecordings array
         DataSource      &upstream;                      // The component producing data to process
         //ManagedBuffer   *savedRecording[BUFFER_SIZE];   // Array of mic buffers that make up a recording
-        uint8_t         *savedRecording[BUFFER_SIZE];   // Array of mic buffers that make up a recording
+        //uint8_t         *savedRecording[BUFFER_SIZE];   // Array of ints from the buffer payload that make up a recording
+        int16_t         *savedRecording[BUFFER_SIZE];   // Array of ints from the buffer payload that make up a recording
+
         Mixer2          &mixer;
 
         /**
